requirements.txt
asgiref==3.2.7
Babel==2.8.0
Django==3.0.5
django-phonenumber-field==4.0.0
phonenumbers==8.12.1
Pillow==7.1.1
pkg-resources==0.0.0
pytz==2019.3
sqlparse==0.3.1
